requirements.txt
streamlit
chromadb
ollama
pymupdf  # for fitz
pytesseract
Pillow
sentence-transformers
torch  # required by sentence-transformers
numpy  # used internally by several packages
